Separate common test functions into new file
